GRIDEDIT-872 Replaced all occurrences of int with size_t

+++ poc/undo/include/AddEdgeTransaction.hpp
@@ -11,32 +11,32 @@ class SimpleMesh;
 class AddEdgeTransaction : public BaseMeshTransaction<AddEdgeTransaction, SimpleMesh>
 {
 public:
-    static TransactionPtr create(SimpleMesh& mesh, const int id, const int start, const int end)
+    static TransactionPtr create(SimpleMesh& mesh, const size_t id, const size_t start, const size_t end)
     {
         return std::make_unique<AddEdgeTransaction>(mesh, id, start, end);
     }
 
-    AddEdgeTransaction(SimpleMesh& mesh, const int id, const int start, const int end) : BaseMeshTransaction<AddEdgeTransaction, SimpleMesh>(mesh), edgeId_(id), start_(start), end_(end) {}
+    AddEdgeTransaction(SimpleMesh& mesh, const size_t id, const size_t start, const size_t end) : BaseMeshTransaction<AddEdgeTransaction, SimpleMesh>(mesh), edgeId_(id), start_(start), end_(end) {}
 
-    int edgeId() const
+    size_t edgeId() const
     {
         return edgeId_;
     }
 
-    int start() const
+    size_t start() const
     {
         return start_;
     }
 
-    int end() const
+    size_t end() const
     {
         return end_;
     }
 
 private:
-    int edgeId_;
-    int start_;
-    int end_;
+    size_t edgeId_;
+    size_t start_;
+    size_t end_;
 };
 
 #endif // ADD_EDGE_TRANSACTION__HPP
